Refactor DashboardPostDetailActivity into extension files
Decomposes the large activity into smaller, responsibility-grouped extension files using Kotlin extension functions. This significantly improves code organization, readability, and maintainability by reducing the activity's overall size and complexity.

Updates application version to 0.4.85.
