+++ app/src/main/java/org/ole/planet/myplanet/lite/dashboard/DashboardPostDetailActivity.kt
@@ -68,43 +68,43 @@ import kotlin.math.min
 internal fun transformCommentMarkdownForDisplay(markdown: String): String = markdown.replace("\n", "  \n")
 
 class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() {
-    private lateinit var recyclerView: RecyclerView
-    private lateinit var loadingView: View
+    internal lateinit var recyclerView: RecyclerView
+    internal lateinit var loadingView: View
 
-    private val repository =
+    internal val repository =
         DashboardNewsRepository(
             client = OkHttpClient.Builder().build(),
             moshi = Moshi.Builder().addLast(KotlinJsonAdapterFactory()).build(),
         )
-    private val actionsRepository = DashboardNewsActionsRepository(AuthDependencies.client, AuthDependencies.moshi, Dispatchers.IO)
-    private val composerRepository =
+    internal val actionsRepository = DashboardNewsActionsRepository(AuthDependencies.client, AuthDependencies.moshi, Dispatchers.IO)
+    internal val composerRepository =
         VoicesComposerRepository(
             client = OkHttpClient.Builder().build(),
             moshi = Moshi.Builder().addLast(KotlinJsonAdapterFactory()).build(),
         )
-    private val httpClient = OkHttpClient.Builder().build()
-    private lateinit var adapter: PostDetailAdapter
-    private lateinit var markwon: Markwon
+    internal val httpClient = OkHttpClient.Builder().build()
+    internal lateinit var adapter: PostDetailAdapter
+    internal lateinit var markwon: Markwon
 
-    private lateinit var replyContainer: View
-    private lateinit var replyInputLayout: TextInputLayout
-    private lateinit var replyInput: EditText
-    private lateinit var replyExpandedContent: View
-    private lateinit var replyPreviewLabel: TextView
-    private lateinit var replyPreview: TextView
-    private lateinit var replyPreviewContainer: LinearLayout
-    private lateinit var replyPreviewImagesRow: View
-    private lateinit var replyPreviewImages: LinearLayout
-    private lateinit var replySendButton: MaterialButton
-    private lateinit var replyActionsRow: View
-    private lateinit var replyMarkdownToolbar: LinearLayout
-    private lateinit var replyingToLabel: TextView
-    private lateinit var backCallback: OnBackPressedCallback
+    internal lateinit var replyContainer: View
+    internal lateinit var replyInputLayout: TextInputLayout
+    internal lateinit var replyInput: EditText
+    internal lateinit var replyExpandedContent: View
+    internal lateinit var replyPreviewLabel: TextView
+    internal lateinit var replyPreview: TextView
+    internal lateinit var replyPreviewContainer: LinearLayout
+    internal lateinit var replyPreviewImagesRow: View
+    internal lateinit var replyPreviewImages: LinearLayout
+    internal lateinit var replySendButton: MaterialButton
+    internal lateinit var replyActionsRow: View
+    internal lateinit var replyMarkdownToolbar: LinearLayout
+    internal lateinit var replyingToLabel: TextView
+    internal lateinit var backCallback: OnBackPressedCallback
 
-    private val pendingReplyImages = LinkedHashMap<String, PendingVoiceImage>()
-    private var replyPendingNewlineIndex: Int? = null
-    private var isHandlingReplyListContinuation = false
-    private val replyImagePickerLauncher =
+    internal val pendingReplyImages = LinkedHashMap<String, PendingVoiceImage>()
+    internal var replyPendingNewlineIndex: Int? = null
+    internal var isHandlingReplyListContinuation = false
+    internal val replyImagePickerLauncher =
         registerForActivityResult(ActivityResultContracts.GetContent()) { uri ->
             uri?.let {
                 lifecycleScope.launch {
@@ -112,7 +112,7 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
                 }
             }
         }
-    private val editVoiceLauncher =
+    internal val editVoiceLauncher =
         registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
             if (result.resultCode == Activity.RESULT_OK) {
                 setResult(Activity.RESULT_OK)
@@ -120,28 +120,28 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
             }
         }
 
-    private var avatarLoader: DashboardAvatarLoader? = null
-    private var imageLoader: DashboardPostImageLoader? = null
-    private var shareHelper: PostShareHelper? = null
-    private var baseUrl: String? = null
-    private var sessionCookie: String? = null
-    private var credentials: StoredCredentials? = null
-    private var serverCode: String? = null
-    private var serverParentCode: String? = null
-    private var selectedTeamId: String? = null
-    private var selectedTeamName: String? = null
-    private var currentUsername: String? = null
-    private var isUserAdmin: Boolean = false
-    private var currentComments: List<PostDetailItem.Comment> = emptyList()
-    private var isEditingComment: Boolean = false
-    private var editingCommentDocument: NewsDocument? = null
-    private var document: NewsDocument? = null
-    private var isPostingReply: Boolean = false
-    private var isReplyComposerExpanded: Boolean = false
-    private var replyContextHandle: String? = null
+    internal var avatarLoader: DashboardAvatarLoader? = null
+    internal var imageLoader: DashboardPostImageLoader? = null
+    internal var shareHelper: PostShareHelper? = null
+    internal var baseUrl: String? = null
+    internal var sessionCookie: String? = null
+    internal var credentials: StoredCredentials? = null
+    internal var serverCode: String? = null
+    internal var serverParentCode: String? = null
+    internal var selectedTeamId: String? = null
+    internal var selectedTeamName: String? = null
+    internal var currentUsername: String? = null
+    internal var isUserAdmin: Boolean = false
+    internal var currentComments: List<PostDetailItem.Comment> = emptyList()
+    internal var isEditingComment: Boolean = false
+    internal var editingCommentDocument: NewsDocument? = null
+    internal var document: NewsDocument? = null
+    internal var isPostingReply: Boolean = false
+    internal var isReplyComposerExpanded: Boolean = false
+    internal var replyContextHandle: String? = null
 
-    private lateinit var headerItem: PostDetailItem.Header
-    private var avatarUpdateListener: AvatarUpdateNotifier.Listener? = null
+    internal lateinit var headerItem: PostDetailItem.Header
+    internal var avatarUpdateListener: AvatarUpdateNotifier.Listener? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -157,7 +157,7 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
         loadInitialData()
     }
 
-    private fun setupToolbar() {
+    internal fun setupToolbar() {
         val toolbar: MaterialToolbar = findViewById(R.id.postDetailToolbar)
         setSupportActionBar(toolbar)
         supportActionBar?.setDisplayHomeAsUpEnabled(true)
@@ -179,136 +179,30 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
         }
     }
 
-    private fun setupViews() {
+    internal fun setupViews() {
         markwon = Markwon.builder(this).build()
         recyclerView = findViewById(R.id.postDetailRecyclerView)
         loadingView = findViewById(R.id.postDetailLoading)
     }
 
-    private fun setupBackNavigation() {
-        backCallback =
-            object : OnBackPressedCallback(true) {
-                override fun handleOnBackPressed() {
-                    if (collapseReplyComposerIfExpanded()) {
-                        return
-                    }
-                    isEnabled = false
-                    onBackPressedDispatcher.onBackPressed()
-                }
-            }
-        onBackPressedDispatcher.addCallback(this, backCallback)
+    internal fun hasAdapter(): Boolean = ::adapter.isInitialized
+
+    @Suppress("unused")
+    private fun applyWrappedFormatting(
+        prefix: String,
+        suffix: String,
+        placeholder: String,
+        placeCursorInsideWhenNoSelection: Boolean = false,
+    ) {
+        applyWrappedFormattingInternal(prefix, suffix, placeholder, placeCursorInsideWhenNoSelection)
     }
 
-    private fun setupReplyComposer() {
-        replyContainer = findViewById(R.id.postDetailReplyContainer)
-        replyInputLayout = findViewById(R.id.dashboardReplyInputLayout)
-        replyInput = findViewById(R.id.dashboardReplyInput)
-        replyExpandedContent = findViewById(R.id.dashboardReplyExpandedContent)
-        replyPreviewLabel = findViewById(R.id.dashboardReplyPreviewLabel)
-        replyPreview = findViewById(R.id.dashboardReplyPreview)
-        replyPreviewContainer = findViewById(R.id.dashboardReplyPreviewContainer)
-        replyPreviewImagesRow = findViewById(R.id.dashboardReplyPreviewImagesRow)
-        replyPreviewImages = findViewById(R.id.dashboardReplyPreviewImages)
-        replySendButton = findViewById(R.id.dashboardReplySendButton)
-        replyActionsRow = findViewById(R.id.dashboardReplyActions)
-        replyMarkdownToolbar = findViewById(R.id.dashboardReplyMarkdownToolbar)
-        replyingToLabel = findViewById(R.id.postDetailReplyingTo)
-
-        setupReplyWindowInsets()
-        setupReplyInputListeners()
-        setupReplyMarkdownToolbar()
-
-        updateReplyPreview(replyPreview, "")
-        setMarkdownToolbarEnabled(false)
-        replySendButton.isEnabled = false
+    @Suppress("unused")
+    private fun updateReplyActionAvailability(text: CharSequence?) {
+        updateReplyActionAvailabilityInternal(text)
     }
 
-    private fun setupReplyWindowInsets() {
-        val baseReplyContainerPaddingBottom = replyContainer.paddingBottom
-
-        ViewCompat.setOnApplyWindowInsetsListener(replyContainer) { view, insets ->
-            val systemBarsBottom = insets.getInsets(WindowInsetsCompat.Type.systemBars()).bottom
-            val imeBottom = insets.getInsets(WindowInsetsCompat.Type.ime()).bottom
-            val bottomInset = max(systemBarsBottom, imeBottom)
-            view.setPadding(
-                view.paddingLeft,
-                view.paddingTop,
-                view.paddingRight,
-                baseReplyContainerPaddingBottom + bottomInset,
-            )
-            insets
-        }
-        ViewCompat.requestApplyInsets(replyContainer)
-    }
-
-    private fun setupReplyInputListeners() {
-        replyInputLayout.helperText = null
-        replyInput.doAfterTextChanged { text ->
-            updateReplyPreview(replyPreview, text?.toString())
-            updateReplyActionAvailability(text)
-        }
-        replyInput.addTextChangedListener(replyListContinuationWatcher)
-        replyInput.setOnFocusChangeListener { _, hasFocus ->
-            if (hasFocus) {
-                expandReplyComposer()
-            }
-        }
-        replyInput.setOnTouchListener { v, event ->
-            if (event.actionMasked == MotionEvent.ACTION_DOWN) {
-                v.parent?.requestDisallowInterceptTouchEvent(true)
-            }
-            if (event.actionMasked == MotionEvent.ACTION_UP || event.actionMasked == MotionEvent.ACTION_CANCEL) {
-                v.parent?.requestDisallowInterceptTouchEvent(false)
-            }
-            false
-        }
-        replyInput.isVerticalScrollBarEnabled = true
-        replyInput.setOnClickListener {
-            expandReplyComposer()
-        }
-        replySendButton.setOnClickListener {
-            val message =
-                replyInput.text
-                    ?.toString()
-                    ?.trim()
-                    .orEmpty()
-            if (isEditingComment) {
-                attemptUpdateComment(message)
-            } else {
-                attemptReply(message)
-            }
-        }
-    }
-
-    private fun setupReplyMarkdownToolbar() {
-        val replyBold: MaterialButton = findViewById(R.id.dashboardReplyMarkdownBold)
-        val replyItalic: MaterialButton = findViewById(R.id.dashboardReplyMarkdownItalic)
-        val replyHeading: MaterialButton = findViewById(R.id.dashboardReplyMarkdownHeading)
-        val replyBullet: MaterialButton = findViewById(R.id.dashboardReplyMarkdownBullet)
-        val replyNumbered: MaterialButton = findViewById(R.id.dashboardReplyMarkdownNumbered)
-        val replyQuote: MaterialButton = findViewById(R.id.dashboardReplyMarkdownQuote)
-        val replyLink: MaterialButton = findViewById(R.id.dashboardReplyMarkdownLink)
-        val replyImage: MaterialButton = findViewById(R.id.dashboardReplyMarkdownImage)
-
-        replyBold.setOnClickListener {
-            applyWrappedFormatting("**", "**", "", placeCursorInsideWhenNoSelection = true)
-        }
-        replyItalic.setOnClickListener {
-            applyWrappedFormatting("*", "*", "", placeCursorInsideWhenNoSelection = true)
-        }
-        replyHeading.setOnClickListener { applyReplyHeadingFormatting() }
-        replyBullet.setOnClickListener { applyLinePrefix("- ") }
-        replyNumbered.setOnClickListener { applyLinePrefix("1. ") }
-        replyQuote.setOnClickListener { applyLinePrefix("> ") }
-        replyLink.setOnClickListener {
-            applyWrappedFormatting("[", "](https://)", "", true)
-        }
-        replyImage.setOnClickListener {
-            handleReplyInsertImageClick()
-        }
-    }
-
-    private fun loadIntentData(): Boolean {
+    internal fun loadIntentData(): Boolean {
         val postId = intent.getStringExtra(EXTRA_POST_ID)
         if (postId.isNullOrBlank()) {
             finish()
@@ -350,976 +244,12 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
         return true
     }
 
-    private fun setupAdapter() {
-        adapter =
-            PostDetailAdapter(
-                markwon,
-                avatarBinder = { imageView, user, hasAvatar ->
-                    val shouldAttemptLoad = hasAvatar || !user.isNullOrBlank()
-                    avatarLoader?.bind(imageView, user, shouldAttemptLoad)
-                },
-                imageBinder = { imageView, path ->
-                    val loader = imageLoader
-                    if (loader != null) {
-                        loader.bind(imageView, path)
-                    } else {
-                        imageView.isVisible = false
-                        imageView.setImageDrawable(null)
-                    }
-                },
-                onImageClicked = { paths, index ->
-                    openImagePreview(paths, index)
-                },
-                onDeleteClicked = {
-                    attemptDeletePost()
-                },
-                onShareClicked = {
-                    shareCurrentPost()
-                },
-                onEditClicked = { header ->
-                    launchEditVoice(header)
-                },
-                onReplyClicked = {
-                    promptReply()
-                },
-                onCommentEditClicked = { comment ->
-                    startEditingComment(comment)
-                },
-                onCommentDeleteClicked = { comment ->
-                    attemptDeleteComment(comment)
-                },
-            )
-
-        recyclerView.layoutManager = LinearLayoutManager(this)
-        recyclerView.adapter = adapter
-        submitItems(currentComments)
-    }
-
-    private fun loadInitialData() {
-        lifecycleScope.launch {
-            initializeSession()
-            val profile = loadCachedProfile()
-            currentUsername = profile?.username
-            isUserAdmin = profile?.isUserAdmin == true
-            refreshHeaderActions()
-            val base = baseUrl
-            if (base.isNullOrBlank()) {
-                Toast
-                    .makeText(
-                        this@DashboardPostDetailActivity,
-                        R.string.dashboard_post_detail_comments_error,
-                        Toast.LENGTH_SHORT,
-                    ).show()
-                updateItems(emptyList())
-                loadingView.isVisible = false
-                return@launch
-            }
-            avatarLoader = DashboardAvatarLoader(base, sessionCookie, credentials, lifecycleScope)
-            avatarUpdateListener =
-                AvatarUpdateNotifier.register(
-                    AvatarUpdateNotifier.Listener { username ->
-                        handleAvatarUpdated(username)
-                    },
-                )
-            imageLoader = DashboardPostImageLoader(base, sessionCookie, lifecycleScope)
-            shareHelper =
-                PostShareHelper(
-                    applicationContext,
-                    { baseUrl },
-                    { sessionCookie },
-                    { serverCode ?: baseUrl?.let { Uri.parse(it).host } },
-                )
-            loadComments(headerItem.id)
-        }
-    }
-
     override fun onSupportNavigateUp(): Boolean {
         onBackPressedDispatcher.onBackPressed()
         return true
     }
 
-    private var cachedProfile: UserProfile? = null
-
-    private fun handleAvatarUpdated(username: String) {
-        if (!::adapter.isInitialized) {
-            return
-        }
-        val positions =
-            adapter.currentList.mapIndexedNotNull { index, item ->
-                val itemUsername =
-                    when (item) {
-                        is PostDetailItem.Header -> item.username
-                        is PostDetailItem.Comment -> item.username
-                    }
-                if (itemUsername?.equals(username, ignoreCase = true) == true) {
-                    index
-                } else {
-                    null
-                }
-            }
-        if (positions.isEmpty()) {
-            return
-        }
-        recyclerView.post {
-            positions.forEach { position ->
-                adapter.notifyItemChanged(position)
-            }
-        }
-    }
-
-    private suspend fun initializeSession() {
-        val context = applicationContext
-        baseUrl = DashboardServerPreferences.getServerBaseUrl(context)
-        val session = ProfileCredentialsStore.getCurrentUserSession(context)
-        credentials = session?.credentials
-        cachedProfile = session?.profile
-        serverCode = DashboardServerPreferences.getServerCode(context)
-        serverParentCode = DashboardServerPreferences.getServerParentCode(context)
-        baseUrl?.let { base ->
-            val authService = AuthDependencies.provideAuthService(context, base)
-            sessionCookie = authService.getStoredToken()
-        }
-    }
-
-    private suspend fun loadComments(postId: String) {
-        val base = baseUrl ?: return
-        loadingView.isVisible = true
-        val result =
-            repository.fetchComments(
-                base,
-                sessionCookie,
-                postId,
-                COMMENTS_LIMIT,
-                serverCode,
-                serverParentCode,
-                selectedTeamName,
-            )
-        result
-            .onSuccess { docs ->
-                val sorted = docs.sortedBy { it.time ?: 0L }
-                val mapped = sorted.mapNotNull { mapToCommentItem(it) }
-                updateItems(mapped)
-            }.onFailure {
-                Toast
-                    .makeText(
-                        this,
-                        R.string.dashboard_post_detail_comments_error,
-                        Toast.LENGTH_SHORT,
-                    ).show()
-                updateItems(emptyList())
-            }
-        loadingView.isVisible = false
-    }
-
-    private fun updateItems(comments: List<PostDetailItem.Comment>) {
-        headerItem =
-            headerItem.copy(
-                commentCount = comments.size,
-                isLoadingComments = false,
-            )
-        currentComments = comments
-        submitItems(currentComments)
-    }
-
-    private fun refreshHeaderActions() {
-        val username = headerItem.username
-        val hasSession = !sessionCookie.isNullOrBlank()
-        val isAuthor = hasSession && !username.isNullOrBlank() && currentUsername?.equals(username, ignoreCase = true) == true
-        val canShare = hasSession
-        val canDelete = (isAuthor || isUserAdmin) && document != null
-        headerItem =
-            headerItem.copy(
-                canReply = hasSession,
-                canEdit = isAuthor,
-                canDelete = canDelete,
-                canShare = canShare,
-            )
-        updateReplyComposerVisibility()
-        submitItems(currentComments)
-    }
-
-    private fun promptReply() {
-        if (!headerItem.canReply || isPostingReply) {
-            return
-        }
-        exitCommentEditMode()
-        expandReplyComposer()
-        replyContainer.isVisible = true
-        replyInput.requestFocus()
-        replyInput.post {
-            val imm = getSystemService(InputMethodManager::class.java)
-            imm?.showSoftInput(replyInput, 0)
-        }
-    }
-
-    private fun startEditingComment(comment: PostDetailItem.Comment) {
-        if (isPostingReply) {
-            return
-        }
-        val doc = comment.document
-        if (doc == null) {
-            Toast.makeText(this, R.string.dashboard_comment_edit_error, Toast.LENGTH_SHORT).show()
-            return
-        }
-        isEditingComment = true
-        editingCommentDocument = doc
-        replyInput.hint = getString(R.string.dashboard_comment_edit_hint)
-        replyInputLayout.helperText = getString(R.string.dashboard_comment_edit_helper)
-        replySendButton.setText(R.string.dashboard_comment_edit_send)
-        replyInput.setText(comment.message.orEmpty())
-        replyInput.setSelection(replyInput.text?.length ?: 0)
-        clearPendingReplyImages()
-        lifecycleScope.launch {
-            loadExistingCommentImages(comment)
-        }
-        expandReplyComposer()
-        replyContainer.isVisible = true
-        replyInput.requestFocus()
-        replyInput.post {
-            val imm = getSystemService(InputMethodManager::class.java)
-            imm?.showSoftInput(replyInput, 0)
-        }
-        updateReplyPreview(replyPreview, replyInput.text?.toString())
-    }
-
-    private fun exitCommentEditMode(clearFields: Boolean = false) {
-        if (!isEditingComment) {
-            if (clearFields) {
-                replyInput.setText("")
-                clearPendingReplyImages()
-                updateReplyPreview(replyPreview, "")
-            }
-            return
-        }
-        isEditingComment = false
-        editingCommentDocument = null
-        replyInput.hint = getString(R.string.dashboard_post_reply_hint)
-        replyInputLayout.helperText = null
-        replySendButton.setText(R.string.dashboard_post_reply_send)
-        if (clearFields) {
-            replyInput.setText("")
-            clearPendingReplyImages()
-            updateReplyPreview(replyPreview, "")
-        }
-    }
-
-    private fun attemptReply(message: String) {
-        if (message.isBlank() && pendingReplyImages.isEmpty()) {
-            Toast.makeText(this, R.string.dashboard_post_reply_empty, Toast.LENGTH_SHORT).show()
-            return
-        }
-        val base = baseUrl
-        val postId = headerItem.id
-        val doc = document
-        if (base.isNullOrBlank() || doc == null) {
-            Toast.makeText(this, R.string.dashboard_post_detail_comments_error, Toast.LENGTH_SHORT).show()
-            return
-        }
-        val credentials = this.credentials
-        if (credentials == null) {
-            Toast.makeText(this, R.string.dashboard_post_reply_missing_credentials, Toast.LENGTH_SHORT).show()
-            return
-        }
-        val cookie = sessionCookie
-        if (cookie.isNullOrBlank()) {
-            Toast.makeText(this, R.string.dashboard_post_detail_comments_error, Toast.LENGTH_SHORT).show()
-            return
-        }
-        if (isPostingReply) {
-            return
-        }
-        setReplyPosting(true)
-        lifecycleScope.launch {
-            val prepared = prepareReplyImagesForPosting(base, credentials, message)
-            val userPayload = buildUserPayload(credentials)
-            val result =
-                composerRepository.createVoice(
-                    VoicesComposerRepository.CreateVoiceParams(
-                        baseUrl = base,
-                        credentials = credentials,
-                        sessionCookie = cookie,
-                        message = prepared.message,
-                        createdOn = doc.createdOn ?: serverCode,
-                        parentCode = doc.parentCode,
-                        replyTo = postId,
-                        images = prepared.images,
-                        labels = emptyList(),
-                        userPayload = userPayload,
-                        teamId = selectedTeamId,
-                        teamName = selectedTeamName,
-                    ),
-                )
-            result
-                .onSuccess {
-                    Toast.makeText(this@DashboardPostDetailActivity, R.string.dashboard_post_reply_success, Toast.LENGTH_SHORT).show()
-                    replyInput.setText("")
-                    clearPendingReplyImages()
-                    updateReplyPreview(replyPreview, "")
-                    hideReplyKeyboard()
-                    collapseReplyComposerIfExpanded()
-                    loadComments(postId)
-                }.onFailure {
-                    Toast.makeText(this@DashboardPostDetailActivity, R.string.dashboard_post_reply_error, Toast.LENGTH_SHORT).show()
-                }
-            setReplyPosting(false)
-        }
-    }
-
-    private fun attemptUpdateComment(message: String) {
-        if (message.isBlank() && pendingReplyImages.isEmpty()) {
-            Toast.makeText(this, R.string.dashboard_post_reply_empty, Toast.LENGTH_SHORT).show()
-            return
-        }
-        val base = baseUrl
-        val doc = editingCommentDocument
-        if (base.isNullOrBlank() || doc == null) {
-            Toast.makeText(this, R.string.dashboard_comment_edit_error, Toast.LENGTH_SHORT).show()
-            exitCommentEditMode(clearFields = true)
-            return
-        }
-        val credentials = this.credentials
-        if (credentials == null) {
-            Toast.makeText(this, R.string.dashboard_post_reply_missing_credentials, Toast.LENGTH_SHORT).show()
-            return
-        }
-        val cookie = sessionCookie
-        if (cookie.isNullOrBlank()) {
-            Toast.makeText(this, R.string.dashboard_comment_edit_error, Toast.LENGTH_SHORT).show()
-            return
-        }
-        if (isPostingReply) {
-            return
-        }
-        setReplyPosting(true)
-        lifecycleScope.launch {
-            val prepared = prepareReplyImagesForPosting(base, credentials, message)
-            val newImages =
-                prepared.images.map { image ->
-                    DashboardNewsRepository.NewsImage(
-                        resourceId = image.resourceId,
-                        filename = image.filename,
-                        markdown = image.markdown,
-                    )
-                }
-            val mergedImages = mergeNewsImages(doc.images, newImages)
-            val result =
-                actionsRepository.updateNews(
-                    baseUrl = base,
-                    sessionCookie = cookie,
-                    document = doc,
-                    message = prepared.message,
-                    images = mergedImages,
-                    teamId = selectedTeamId,
-                    teamName = selectedTeamName,
-                )
-            result
-                .onSuccess {
-                    Toast
-                        .makeText(this@DashboardPostDetailActivity, R.string.dashboard_comment_edit_success, Toast.LENGTH_SHORT)
-                        .show()
-                    exitCommentEditMode(clearFields = true)
-                    hideReplyKeyboard()
-                    collapseReplyComposerIfExpanded()
-                    loadComments(headerItem.id)
-                }.onFailure {
-                    Toast
-                        .makeText(this@DashboardPostDetailActivity, R.string.dashboard_comment_edit_error, Toast.LENGTH_SHORT)
-                        .show()
-                }
-            setReplyPosting(false)
-        }
-    }
-
-    private fun updateReplyPreview(
-        preview: TextView,
-        message: String?,
-    ) {
-        val content = message?.takeIf { it.isNotBlank() }
-        val previewText = content ?: getString(R.string.dashboard_post_reply_preview_placeholder)
-        val transformed = transformReplyMarkdownForPreview(previewText)
-        markwon.setMarkdown(preview, transformed)
-        updateReplyPreviewImages()
-    }
-
-    private fun mergeNewsImages(
-        existingImages: List<DashboardNewsRepository.NewsImage>?,
-        newImages: List<DashboardNewsRepository.NewsImage>,
-    ): List<DashboardNewsRepository.NewsImage> {
-        if (existingImages.isNullOrEmpty() && newImages.isEmpty()) {
-            return emptyList()
-        }
-        val merged = LinkedHashMap<String, DashboardNewsRepository.NewsImage>()
-        existingImages?.forEach { image ->
-            val key = image.markdown ?: buildResourcePath(image.resourceId, image.filename)
-            if (!key.isNullOrBlank() && !merged.containsKey(key)) {
-                merged[key] = image
-            }
-        }
-        newImages.forEach { image ->
-            val key = image.markdown ?: buildResourcePath(image.resourceId, image.filename)
-            if (!key.isNullOrBlank() && !merged.containsKey(key)) {
-                merged[key] = image
-            }
-        }
-        return merged.values.toList()
-    }
-
-    private fun updateReplyPreviewImages() {
-        val images = pendingReplyImages.values.toList()
-        if (!isReplyComposerExpanded || images.isEmpty()) {
-            replyPreviewImages.removeAllViews()
-            replyPreviewImagesRow.isVisible = false
-            return
-        }
-
-        replyPreviewImages.removeAllViews()
-        val size = resources.getDimensionPixelSize(R.dimen.dashboard_reply_preview_image_size)
-        val spacing = resources.getDimensionPixelSize(R.dimen.dashboard_reply_preview_image_spacing)
-        images.forEachIndexed { index, pending ->
-            val imageView = AppCompatImageView(this)
-            val params = LinearLayout.LayoutParams(size, size)
-            if (index < images.lastIndex) {
-                params.marginEnd = spacing
-            }
-            imageView.layoutParams = params
-            imageView.scaleType = ImageView.ScaleType.CENTER_CROP
-            imageView.clipToOutline = true
-            imageView.setImageURI(Uri.fromFile(pending.file))
-            replyPreviewImages.addView(imageView)
-        }
-        replyPreviewImagesRow.isVisible = true
-    }
-
-    private fun updateReplyComposerVisibility() {
-        val canReply = headerItem.canReply
-        replyInputLayout.isEnabled = canReply && !isPostingReply
-        applyReplyExpansionState()
-        updateReplyActionAvailability(replyInput.text)
-    }
-
-    private fun updateReplyActionAvailability(text: CharSequence?) {
-        val hasContent = !text.isNullOrBlank() || pendingReplyImages.isNotEmpty()
-        val canSend = (headerItem.canReply || isEditingComment) && !isPostingReply && hasContent && isReplyComposerExpanded
-        replySendButton.isEnabled = canSend
-    }
-
-    private fun clearPendingReplyImages() {
-        val filesToDelete = pendingReplyImages.values.map { it.file }.toList()
-        pendingReplyImages.clear()
-
-        @OptIn(kotlinx.coroutines.DelicateCoroutinesApi::class)
-        GlobalScope.launch(Dispatchers.IO) {
-            filesToDelete.forEach { file ->
-                if (file.exists()) {
-                    file.delete()
-                }
-            }
-        }
-        updateReplyPreviewImages()
-        updateReplyActionAvailability(replyInput.text)
-    }
-
-    private fun setReplyPosting(posting: Boolean) {
-        isPostingReply = posting
-        updateReplyComposerVisibility()
-    }
-
-    private fun setMarkdownToolbarEnabled(enabled: Boolean) {
-        replyMarkdownToolbar.isEnabled = enabled
-        for (index in 0 until replyMarkdownToolbar.childCount) {
-            replyMarkdownToolbar.getChildAt(index)?.isEnabled = enabled
-        }
-    }
-
-    private fun handleReplyInsertImageClick() {
-        launchReplyImagePicker()
-    }
-
-    private fun launchReplyImagePicker() {
-        replyImagePickerLauncher.launch("image/*")
-    }
-
-    private suspend fun handleReplyImageSelection(uri: Uri) {
-        val pendingResult =
-            withContext(Dispatchers.IO) {
-                runCatching { VoiceImageFactory.createPendingVoiceImage(uri, contentResolver, cacheDir, ::generatePendingImageId) }
-            }
-        pendingResult
-            .onSuccess { pending ->
-                pendingReplyImages[pending.id] = pending
-                updateReplyPreview(replyPreview, replyInput.text?.toString())
-                updateReplyActionAvailability(replyInput.text)
-            }.onFailure {
-                Toast.makeText(this, R.string.create_voice_image_processing_error, Toast.LENGTH_SHORT).show()
-            }
-    }
-
-    private fun collapseReplyComposerIfExpanded(): Boolean {
-        if (!isReplyComposerExpanded) {
-            return false
-        }
-        isReplyComposerExpanded = false
-        exitCommentEditMode(clearFields = true)
-        hideReplyKeyboard()
-        replyInput.clearFocus()
-        applyReplyExpansionState()
-        updateReplyActionAvailability(replyInput.text)
-        return true
-    }
-
-    private fun expandReplyComposer() {
-        if (isReplyComposerExpanded) {
-            return
-        }
-        isReplyComposerExpanded = true
-        applyReplyExpansionState()
-    }
-
-    private fun applyReplyExpansionState() {
-        val canReply = headerItem.canReply
-        val expanded = canReply && isReplyComposerExpanded
-        replyContainer.isVisible = canReply
-        val minLines = if (expanded) EXPANDED_REPLY_MIN_LINES else COLLAPSED_REPLY_MIN_LINES
-        replyInput.setMinLines(minLines)
-        replyExpandedContent.isVisible = expanded
-        replyMarkdownToolbar.isVisible = expanded
-        replyPreviewContainer.isVisible = expanded
-        replyPreviewLabel.isVisible = expanded
-        replyPreview.isVisible = expanded
-        replyActionsRow.isVisible = expanded
-        updateReplyPreviewImages()
-        replyInputLayout.helperText = null
-        setMarkdownToolbarEnabled(expanded && !isPostingReply)
-        updateReplyingToVisibility(expanded)
-    }
-
-    private fun applyWrappedFormatting(
-        prefix: String,
-        suffix: String,
-        placeholder: String,
-        placeCursorInsideWhenNoSelection: Boolean = false,
-    ) {
-        if (!headerItem.canReply || isPostingReply) {
-            return
-        }
-        val editable = replyInput.text ?: return
-        val start = max(replyInput.selectionStart, 0)
-        val end = max(replyInput.selectionEnd, 0)
-        val selectionStart = min(start, end)
-        val selectionEnd = max(start, end)
-        val selected = editable.substring(selectionStart, selectionEnd)
-        val replacement =
-            if (selectionStart == selectionEnd) {
-                val defaultText = placeholder
-                "$prefix$defaultText$suffix"
-            } else {
-                "$prefix$selected$suffix"
-            }
-        editable.replace(selectionStart, selectionEnd, replacement)
-        val newCursor =
-            if (selectionStart == selectionEnd && placeCursorInsideWhenNoSelection) {
-                selectionStart + prefix.length
-            } else {
-                selectionStart + replacement.length
-            }
-        val boundedCursor = min(max(newCursor, 0), editable.length)
-        replyInput.setSelection(boundedCursor)
-    }
-
-    private fun applyReplyHeadingFormatting() {
-        if (!headerItem.canReply || isPostingReply) {
-            return
-        }
-        val editText = replyInput
-        val editable = editText.text ?: return
-        val cursor = max(0, editText.selectionStart)
-        val lineStart = findReplyLineStart(editable, cursor)
-        val lineLength = editable.length
-        var prefixEnd = lineStart
-        var currentHashes = 0
-        while (prefixEnd < lineLength && editable[prefixEnd] == '#') {
-            currentHashes++
-            prefixEnd++
-        }
-        while (prefixEnd < lineLength && editable[prefixEnd].isWhitespace() && editable[prefixEnd] != '\n') {
-            prefixEnd++
-        }
-        val newHeadingLevel =
-            if (currentHashes == 0) {
-                1
-            } else {
-                (currentHashes % MAX_HEADING_LEVEL) + 1
-            }
-        val replacement =
-            buildString {
-                repeat(newHeadingLevel) { append('#') }
-                append(' ')
-            }
-        editable.replace(lineStart, prefixEnd, replacement)
-        val selection = (lineStart + replacement.length).coerceAtMost(editable.length)
-        editText.setSelection(selection)
-    }
-
-    private fun applyLinePrefix(prefix: String) {
-        if (!headerItem.canReply || isPostingReply) {
-            return
-        }
-        val editable = replyInput.text ?: return
-        val selectionStart = max(replyInput.selectionStart, 0)
-        val lastLineBreak = editable.lastIndexOf('\n', selectionStart - 1)
-        val lineStart = if (lastLineBreak == -1) 0 else lastLineBreak + 1
-        editable.insert(lineStart, prefix)
-        val cursor = min(selectionStart + prefix.length, editable.length)
-        replyInput.setSelection(cursor)
-    }
-
-    private val replyListContinuationWatcher =
-        object : TextWatcher {
-            override fun beforeTextChanged(
-                s: CharSequence?,
-                start: Int,
-                count: Int,
-                after: Int,
-            ) = Unit
-
-            override fun onTextChanged(
-                s: CharSequence?,
-                start: Int,
-                before: Int,
-                count: Int,
-            ) {
-                if (isHandlingReplyListContinuation) {
-                    return
-                }
-                if (count <= 0 || s == null) {
-                    return
-                }
-                val inserted = s.subSequence(start, start + count)
-                val newlineOffset = inserted.lastIndexOf('\n')
-                if (newlineOffset >= 0) {
-                    replyPendingNewlineIndex = start + newlineOffset
-                }
-            }
-
-            override fun afterTextChanged(s: Editable?) {
-                if (isHandlingReplyListContinuation) {
-                    return
-                }
-                val newlineIndex = replyPendingNewlineIndex ?: return
-                replyPendingNewlineIndex = null
-                s ?: return
-                handleReplyListContinuation(s, newlineIndex)
-            }
-        }
-
-    private fun handleReplyListContinuation(
-        editable: Editable,
-        newlineIndex: Int,
-    ) {
-        if (newlineIndex <= 0 || newlineIndex > editable.length) {
-            return
-        }
-        val previousLineStart = findReplyLineStart(editable, newlineIndex)
-        val previousLine = editable.subSequence(previousLineStart, newlineIndex).toString()
-        if (previousLine.isEmpty()) {
-            return
-        }
-        val indentLength = findIndentLength(previousLine)
-        val indent = previousLine.substring(0, indentLength)
-        val contentAfterIndent = previousLine.substring(indentLength)
-        if (contentAfterIndent.isBlank()) {
-            return
-        }
-
-        if (contentAfterIndent.startsWith("- ") || contentAfterIndent.startsWith("* ")) {
-            val marker = contentAfterIndent.substring(0, 2)
-            val hasText = contentAfterIndent.substring(2).isNotBlank()
-            if (hasText) {
-                insertReplyListPrefix(editable, newlineIndex, "$indent$marker")
-            } else {
-                removeReplyListPrefix(editable, previousLineStart + indentLength, marker.length)
-            }
-            return
-        }
-
-        val numberMatch = NUMBERED_LIST_REGEX.matchEntire(contentAfterIndent)
-        if (numberMatch != null) {
-            val number = numberMatch.groupValues.getOrNull(1)?.toIntOrNull() ?: return
-            val remainder = numberMatch.groupValues.getOrNull(2).orEmpty()
-            val markerLength = numberMatch.groupValues[1].length + 2
-            if (remainder.isBlank()) {
-                removeReplyListPrefix(editable, previousLineStart + indentLength, markerLength)
-            } else {
-                val nextMarker = "$indent${number + 1}. "
-                insertReplyListPrefix(editable, newlineIndex, nextMarker)
-            }
-        }
-    }
-
-    private fun insertReplyListPrefix(
-        editable: Editable,
-        newlineIndex: Int,
-        prefix: String,
-    ) {
-        val insertPosition = (newlineIndex + 1).coerceAtMost(editable.length)
-        isHandlingReplyListContinuation = true
-        editable.insert(insertPosition, prefix)
-        replyInput.setSelection((insertPosition + prefix.length).coerceAtMost(editable.length))
-        isHandlingReplyListContinuation = false
-    }
-
-    private fun removeReplyListPrefix(
-        editable: Editable,
-        start: Int,
-        markerLength: Int,
-    ) {
-        val end = (start + markerLength).coerceAtMost(editable.length)
-        isHandlingReplyListContinuation = true
-        editable.delete(start, end)
-        replyInput.setSelection(start.coerceAtMost(editable.length))
-        isHandlingReplyListContinuation = false
-    }
-
-    private fun findReplyLineStart(
-        editable: Editable,
-        index: Int,
-    ): Int {
-        val boundedIndex = index.coerceIn(0, editable.length)
-        for (i in boundedIndex - 1 downTo 0) {
-            if (editable[i] == '\n') {
-                return i + 1
-            }
-        }
-        return 0
-    }
-
-    private fun findIndentLength(line: String): Int {
-        for (i in line.indices) {
-            if (!line[i].isWhitespace()) {
-                return i
-            }
-        }
-        return line.length
-    }
-
-    private fun hideReplyKeyboard() {
-        val imm = getSystemService(InputMethodManager::class.java)
-        imm?.hideSoftInputFromWindow(replyInput.windowToken, 0)
-    }
-
-    private fun updateReplyingToLabel(username: String?) {
-        replyContextHandle = username?.takeIf { it.isNotBlank() }
-        updateReplyingToVisibility(isReplyComposerExpanded && headerItem.canReply)
-    }
-
-    private fun updateReplyingToVisibility(expanded: Boolean) {
-        val handle = replyContextHandle
-        if (expanded && handle != null) {
-            replyingToLabel.text = getString(R.string.dashboard_post_replying_to, handle)
-            replyingToLabel.isVisible = true
-        } else {
-            replyingToLabel.isVisible = false
-        }
-    }
-
-    private suspend fun prepareReplyImagesForPosting(
-        baseUrl: String,
-        credentials: StoredCredentials,
-        originalMessage: String,
-    ): PreparedVoicePost {
-        if (pendingReplyImages.isEmpty()) {
-            return PreparedVoicePost(originalMessage, emptyList())
-        }
-        val context = buildReplyImageResourceContext(credentials)
-
-        val uploads = pendingReplyImages.values.filter { it.resourceId == null }
-        val uploadResults =
-            coroutineScope {
-                uploads
-                    .map { pending ->
-                        async {
-                            val markdown = ensureReplyImageUpload(baseUrl, credentials, context, pending)
-                            pending to markdown
-                        }
-                    }.awaitAll()
-            }
-
-        val preparedImages = mutableListOf<VoicesComposerRepository.ImagePayload>()
-
-        for ((pending, markdown) in uploadResults) {
-            val resourceId = pending.resourceId
-            if (resourceId != null) {
-                preparedImages +=
-                    VoicesComposerRepository.ImagePayload(
-                        resourceId = resourceId,
-                        filename = pending.fileName,
-                        markdown = markdown,
-                    )
-            }
-        }
-        return PreparedVoicePost(originalMessage, preparedImages)
-    }
-
-    private suspend fun ensureReplyImageUpload(
-        baseUrl: String,
-        credentials: StoredCredentials,
-        context: VoiceImageResourceContext,
-        pending: PendingVoiceImage,
-    ): String {
-        pending.uploadedMarkdown?.let { return it }
-        val existingResourceId = pending.resourceId
-        if (existingResourceId != null) {
-            val markdown = "![](resources/${existingResourceId.trim()}/${pending.fileName.trim()})"
-            pending.uploadedMarkdown = markdown
-            return markdown
-        }
-
-        val metadata = VoicesComposerRepository.ResourceMetadataRequest.fromContext(context, pending.fileName)
-        val creationResponse = composerRepository.createResourceDocument(baseUrl, credentials, metadata)
-        pending.resourceId = creationResponse.id
-        pending.resourceRevision = creationResponse.revision
-        val uploadResponse =
-            composerRepository.uploadResourceBinary(
-                baseUrl,
-                credentials,
-                creationResponse.id,
-                pending.fileName,
-                creationResponse.revision,
-                pending.jpegBytes,
-            )
-        val resolvedResourceId = uploadResponse.resourceId ?: creationResponse.id
-        val resolvedFileName = uploadResponse.filename ?: pending.fileName
-        pending.resourceId = resolvedResourceId
-        pending.resourceRevision = uploadResponse.revision ?: creationResponse.revision
-        val relativeMarkdown =
-            uploadResponse.markdown
-                ?: "![](resources/${resolvedResourceId.trim()}/${resolvedFileName.trim()})"
-        pending.uploadedMarkdown = relativeMarkdown
-        return relativeMarkdown
-    }
-
-    private fun transformReplyMarkdownForPreview(markdown: String): String {
-        var processed = markdown.replace("\n", "  \n")
-        val base = baseUrl?.trim()?.trimEnd('/')?.takeIf { it.isNotEmpty() }
-
-        if (!base.isNullOrEmpty()) {
-            processed =
-                RESOURCES_MARKDOWN_REGEX.replace(processed) { matchResult ->
-                    val path =
-                        matchResult.groupValues
-                            .getOrNull(1)
-                            ?.trim()
-                            .orEmpty()
-                    val absolute = "$base/db/$path"
-                    "![]($absolute)"
-                }
-        }
-
-        val pendingByFileName = pendingReplyImages.values.associateBy { it.fileName }
-        if (pendingByFileName.isNotEmpty()) {
-            val globalPattern = Regex("(!\\[[^]]*]\\()(.*?)(\\))")
-            processed =
-                globalPattern.replace(processed) { matchResult ->
-                    val path = matchResult.groupValues.getOrNull(2).orEmpty()
-                    val pending = pendingByFileName[path]
-                    if (pending != null) {
-                        val prefix = matchResult.groupValues.getOrNull(1).orEmpty()
-                        val suffix = matchResult.groupValues.getOrNull(3).orEmpty()
-                        "$prefix${pending.file.toURI()}$suffix"
-                    } else {
-                        matchResult.value
-                    }
-                }
-        }
-        return processed
-    }
-
-    private suspend fun loadExistingCommentImages(comment: PostDetailItem.Comment) {
-        val base = baseUrl ?: return
-        if (comment.imagePaths.isEmpty()) {
-            return
-        }
-        val loaded =
-            coroutineScope {
-                comment.imagePaths
-                    .map { path ->
-                        async(Dispatchers.IO) {
-                            VoiceImageFetcher.fetchExistingImage(httpClient, cacheDir, sessionCookie, base, path, {
-                                VoiceImageFactory.generateImageFileName()
-                            }, { generatePendingImageId(it) })
-                        }
-                    }.awaitAll()
-                    .filterNotNull()
-                    .toMutableList()
-            }
-        if (loaded.isEmpty()) {
-            return
-        }
-        loaded.forEach { pending ->
-            pendingReplyImages[pending.id] = pending
-        }
-        updateReplyPreview(replyPreview, replyInput.text?.toString())
-    }
-
-    private suspend fun buildReplyImageResourceContext(credentials: StoredCredentials): VoiceImageResourceContext {
-        val preferences = SecurePreferencesProvider.getServerPreferences(applicationContext)
-        val androidId = preferences.getString(KEY_DEVICE_ANDROID_ID, null)?.takeIf { it.isNotBlank() }
-        val customDeviceName = preferences.getString(KEY_DEVICE_CUSTOM_DEVICE_NAME, null)?.takeIf { it.isNotBlank() }
-        val storedServerCode = preferences.getString(KEY_SERVER_CODE, null)?.takeIf { it.isNotBlank() }
-        val storedParentCode = preferences.getString(KEY_SERVER_PARENT_CODE, null)?.takeIf { it.isNotBlank() }
-        val profile = withContext(Dispatchers.IO) { loadCachedProfile() }
-        val parsedCodes = parseCodesFromProfile(profile?.rawDocument)
-        val resolvedResideOn =
-            serverCode?.takeIf { it.isNotBlank() }
-                ?: storedServerCode
-                ?: parsedCodes?.planetCode
-        val resolvedParent = storedParentCode ?: parsedCodes?.parentCode
-        return VoiceImageResourceContext(
-            username = credentials.username,
-            resideOn = resolvedResideOn,
-            sourcePlanet = resolvedParent,
-            androidId = androidId,
-            deviceName =
-                org.ole.planet.myplanet.lite.util.DeviceUtils
-                    .getDeviceName(),
-            customDeviceName = customDeviceName,
-        )
-    }
-
-    private fun parseCodesFromProfile(rawDocument: String?): ProfileCodes? {
-        if (rawDocument.isNullOrBlank()) {
-            return null
-        }
-        return runCatching {
-            val json = org.json.JSONObject(rawDocument)
-            val planetCode = json.optString("planetCode").takeIf { it.isNotBlank() }
-            val parentCode = json.optString("parentCode").takeIf { it.isNotBlank() }
-            ProfileCodes(planetCode, parentCode)
-        }.getOrNull()
-    }
-
-    private suspend fun loadCachedProfile(): UserProfile? {
-        val existing = cachedProfile
-        if (existing != null) {
-            return existing
-        }
-        val profile =
-            withContext(Dispatchers.IO) {
-                UserProfileDatabase.getInstance(applicationContext).getProfile()
-        }
-        cachedProfile = profile
-        return profile
-    }
-
-    private fun generatePendingImageId(baseName: String): String {
-        var candidate = baseName
-        var counter = 1
-        while (pendingReplyImages.containsKey(candidate)) {
-            candidate = "${baseName}_$counter"
-            counter++
-        }
-        return candidate
-    }
+    internal var cachedProfile: UserProfile? = null
 
     override fun onDestroy() {
         super.onDestroy()
@@ -1331,262 +261,6 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
         AvatarUpdateNotifier.unregister(avatarUpdateListener)
         avatarUpdateListener = null
         clearPendingReplyImages()
-    }
-
-    private suspend fun buildUserPayload(credentials: StoredCredentials): VoicesComposerRepository.UserPayload {
-        val profile = loadCachedProfile()
-        val planetCode = serverCode?.takeIf { it.isNotBlank() } ?: document?.createdOn
-        return VoicesComposerRepository.UserPayload(
-            id = "org.couchdb.user:${credentials.username}",
-            name = profile?.username ?: credentials.username,
-            firstName = profile?.firstName,
-            middleName = profile?.middleName,
-            lastName = profile?.lastName,
-            email = profile?.email,
-            language = profile?.language,
-            phoneNumber = profile?.phoneNumber,
-            planetCode = planetCode,
-            parentCode = document?.parentCode,
-            roles = null,
-            joinDate = null,
-            attachments = null,
-        )
-    }
-
-    private fun shareCurrentPost() {
-        val helper = shareHelper ?: return
-        val header = headerItem
-        lifecycleScope.launch {
-            helper.sharePost(header.id, header.author, header.message, header.imagePaths)
-        }
-    }
-
-    private fun attemptDeletePost() {
-        val base = baseUrl
-        val doc = document
-        if (base.isNullOrBlank() || doc == null) {
-            Toast.makeText(this, R.string.dashboard_post_delete_error, Toast.LENGTH_SHORT).show()
-            return
-        }
-        val cookie = sessionCookie
-        if (cookie.isNullOrBlank()) {
-            Toast.makeText(this, R.string.dashboard_post_delete_error, Toast.LENGTH_SHORT).show()
-            return
-        }
-        lifecycleScope.launch {
-            val result =
-                actionsRepository.deleteNews(
-                    base,
-                    cookie,
-                    doc,
-                    teamId = selectedTeamId,
-                    teamName = selectedTeamName,
-                )
-            result
-                .onSuccess {
-                    Toast
-                        .makeText(this@DashboardPostDetailActivity, R.string.dashboard_post_delete_success, Toast.LENGTH_SHORT)
-                        .show()
-                    val deletedIntent = Intent().putExtra(EXTRA_DELETED_POST_ID, doc.id)
-                    setResult(RESULT_OK, deletedIntent)
-                    finish()
-                }.onFailure {
-                    Toast
-                        .makeText(this@DashboardPostDetailActivity, R.string.dashboard_post_delete_error, Toast.LENGTH_SHORT)
-                        .show()
-                }
-        }
-    }
-
-    private fun attemptDeleteComment(comment: PostDetailItem.Comment) {
-        val base = baseUrl
-        val doc = comment.document
-        if (base.isNullOrBlank() || doc == null) {
-            Toast.makeText(this, R.string.dashboard_comment_delete_error, Toast.LENGTH_SHORT).show()
-            return
-        }
-        val cookie = sessionCookie
-        if (cookie.isNullOrBlank()) {
-            Toast.makeText(this, R.string.dashboard_comment_delete_error, Toast.LENGTH_SHORT).show()
-            return
-        }
-        lifecycleScope.launch {
-            val result =
-                actionsRepository.deleteNews(
-                    base,
-                    cookie,
-                    doc,
-                    teamId = selectedTeamId,
-                    teamName = selectedTeamName,
-                )
-            result
-                .onSuccess {
-                    Toast
-                        .makeText(this@DashboardPostDetailActivity, R.string.dashboard_comment_delete_success, Toast.LENGTH_SHORT)
-                        .show()
-                    currentComments = currentComments.filterNot { it.id == comment.id }
-                    updateItems(currentComments)
-                }.onFailure {
-                    Toast
-                        .makeText(this@DashboardPostDetailActivity, R.string.dashboard_comment_delete_error, Toast.LENGTH_SHORT)
-                        .show()
-                }
-        }
-    }
-
-    private fun submitItems(comments: List<PostDetailItem.Comment>) {
-        val newItems = ArrayList<PostDetailItem>(1 + comments.size)
-        newItems.add(headerItem)
-        newItems.addAll(comments)
-        adapter.submitList(newItems)
-    }
-
-    private fun mapToCommentItem(document: NewsDocument): PostDetailItem.Comment? {
-        val id = document.id ?: return null
-        val username = document.user?.name?.takeIf { it.isNotBlank() }
-        val displayName =
-            document.user?.let { user ->
-                val parts =
-                    listOfNotNull(user.firstName, user.middleName, user.lastName)
-                        .map { it.trim() }
-                        .filter { it.isNotEmpty() }
-                when {
-                    parts.isNotEmpty() -> TextUtils.join(" ", parts)
-                    !username.isNullOrEmpty() -> username
-                    else -> getString(R.string.dashboard_profile_name_placeholder)
-                }
-            } ?: (username ?: getString(R.string.dashboard_profile_name_placeholder))
-        val message = document.message?.takeUnless { it.isNullOrBlank() }
-        val imagePaths = mapCommentImages(document)
-        val hasAvatar = document.user?.attachments?.isNullOrEmpty() == false
-        val timestamp = document.time ?: 0L
-        val hasSession = !sessionCookie.isNullOrBlank()
-        val isAuthor = hasSession && !username.isNullOrBlank() && currentUsername?.equals(username, ignoreCase = true) == true
-        val canDelete = hasSession && (isAuthor || isUserAdmin)
-        return PostDetailItem.Comment(
-            id = id,
-            author = displayName,
-            username = username,
-            hasAvatar = hasAvatar,
-            message = message,
-            imagePaths = imagePaths,
-            timestamp = timestamp,
-            canEdit = isAuthor,
-            canDelete = canDelete,
-            document = document,
-        )
-    }
-
-    private fun mapCommentImages(document: NewsDocument): List<String> {
-        val fromImages =
-            document.images
-                ?.mapNotNull { image ->
-                    extractImagePath(image.markdown)
-                        ?: buildResourcePath(image.resourceId, image.filename)
-                }?.filter { it.isNotBlank() }
-                .orEmpty()
-        val fromMessage = collectImagePaths(document.message)
-        return mergeImagePaths(fromImages + fromMessage)
-    }
-
-    private fun collectImagePaths(markdown: String?): List<String> {
-        if (markdown.isNullOrBlank()) {
-            return emptyList()
-        }
-        return IMAGE_MARKDOWN_REGEX
-            .findAll(markdown)
-            .mapNotNull { match ->
-                match.groupValues
-                    .getOrNull(2)
-                    ?.trim()
-                    ?.takeIf { it.isNotEmpty() }
-            }.toList()
-    }
-
-    private fun mergeImagePaths(paths: List<String>): List<String> {
-        if (paths.isEmpty()) {
-            return emptyList()
-        }
-        val seen = LinkedHashSet<String>()
-        val merged = mutableListOf<String>()
-        for (path in paths) {
-            val normalized = normalizeImagePath(path)
-            if (seen.add(normalized)) {
-                merged += path
-            }
-        }
-        return merged
-    }
-
-    private fun normalizeImagePath(path: String): String {
-        val extracted = extractImagePath(path) ?: path
-        val trimmed = extracted.trim()
-        val resourcesMatch = RESOURCES_PATH_REGEX.find(trimmed)
-        val reduced =
-            if (resourcesMatch != null) {
-                resourcesMatch.value
-            } else {
-                trimmed
-                    .trimStart('/')
-                    .removePrefix("db/")
-                    .trimStart('/')
-            }
-        return reduced.lowercase(Locale.US)
-    }
-
-    private fun extractImagePath(markdown: String?): String? {
-        if (markdown.isNullOrBlank()) {
-            return null
-        }
-        val match = IMAGE_MARKDOWN_REGEX.find(markdown)
-        return match
-            ?.groupValues
-            ?.getOrNull(2)
-            ?.trim()
-            ?.takeIf { it.isNotEmpty() }
-    }
-
-    private fun buildResourcePath(
-        resourceId: String?,
-        filename: String?,
-    ): String? {
-        val id = resourceId?.trim().takeUnless { it.isNullOrEmpty() }
-        val name = filename?.trim().takeUnless { it.isNullOrEmpty() }
-        if (id == null || name == null) {
-            return null
-        }
-        return "resources/$id/$name"
-    }
-
-    private fun openImagePreview(
-        imagePaths: List<String>,
-        startIndex: Int,
-    ) {
-        if (imagePaths.isEmpty()) {
-            return
-        }
-        val intent = Intent(this, DashboardImagePreviewActivity::class.java)
-        intent.putStringArrayListExtra(
-            DashboardImagePreviewActivity.EXTRA_IMAGE_PATHS,
-            ArrayList(imagePaths),
-        )
-        intent.putExtra(DashboardImagePreviewActivity.EXTRA_START_INDEX, startIndex)
-        startActivity(intent)
-    }
-
-    private fun launchEditVoice(item: PostDetailItem.Header) {
-        val intent = Intent(this, CreateVoiceActivity::class.java)
-        intent.putExtra(CreateVoiceActivity.EXTRA_IS_EDIT_MODE, true)
-        intent.putExtra(CreateVoiceActivity.EXTRA_EDIT_POST_ID, item.id)
-        intent.putExtra(CreateVoiceActivity.EXTRA_EDIT_INITIAL_MESSAGE, item.message)
-        intent.putStringArrayListExtra(
-            CreateVoiceActivity.EXTRA_EDIT_INITIAL_IMAGE_PATHS,
-            ArrayList(item.imagePaths),
-        )
-        document?.let { intent.putExtra(CreateVoiceActivity.EXTRA_EDIT_DOCUMENT, it) }
-        selectedTeamId?.let { intent.putExtra(CreateVoiceActivity.EXTRA_TARGET_TEAM_ID, it) }
-        selectedTeamName?.let { intent.putExtra(CreateVoiceActivity.EXTRA_TARGET_TEAM_NAME, it) }
-        editVoiceLauncher.launch(intent)
     }
 
     companion object {
@@ -1603,27 +277,27 @@ class DashboardPostDetailActivity : org.ole.planet.myplanet.lite.BaseActivity() 
         const val EXTRA_TEAM_NAME = "extra_team_name"
         const val EXTRA_DELETED_POST_ID = "extra_deleted_post_id"
 
-        private const val COMMENTS_LIMIT = 50
-        private const val MINUTE_MILLIS = 60_000L
-        private const val HOUR_MILLIS = 60 * MINUTE_MILLIS
-        private const val DAY_MILLIS = 24 * HOUR_MILLIS
-        private const val MONTH_MILLIS = 30 * DAY_MILLIS
-        private const val YEAR_MILLIS = 12 * MONTH_MILLIS
-        private const val PRIVATE_FOR_COMMUNITY = "community"
-        private const val JPEG_QUALITY = 85
-        private const val MAX_IMAGE_DIMENSION = 1280
-        private const val PREFS_NAME = "PREFS"
-        private const val KEY_DEVICE_ANDROID_ID = "device_android_id"
-        private const val KEY_DEVICE_CUSTOM_DEVICE_NAME = "device_custom_device_name"
-        private const val KEY_SERVER_PARENT_CODE = "server_parent_code"
-        private const val KEY_SERVER_CODE = "server_code"
-        private const val COLLAPSED_REPLY_MIN_LINES = 1
-        private const val EXPANDED_REPLY_MIN_LINES = 3
-        private const val MAX_HEADING_LEVEL = 6
+        internal const val COMMENTS_LIMIT = 50
+        internal const val MINUTE_MILLIS = 60_000L
+        internal const val HOUR_MILLIS = 60 * MINUTE_MILLIS
+        internal const val DAY_MILLIS = 24 * HOUR_MILLIS
+        internal const val MONTH_MILLIS = 30 * DAY_MILLIS
+        internal const val YEAR_MILLIS = 12 * MONTH_MILLIS
+        internal const val PRIVATE_FOR_COMMUNITY = "community"
+        internal const val JPEG_QUALITY = 85
+        internal const val MAX_IMAGE_DIMENSION = 1280
+        internal const val PREFS_NAME = "PREFS"
+        internal const val KEY_DEVICE_ANDROID_ID = "device_android_id"
+        internal const val KEY_DEVICE_CUSTOM_DEVICE_NAME = "device_custom_device_name"
+        internal const val KEY_SERVER_PARENT_CODE = "server_parent_code"
+        internal const val KEY_SERVER_CODE = "server_code"
+        internal const val COLLAPSED_REPLY_MIN_LINES = 1
+        internal const val EXPANDED_REPLY_MIN_LINES = 3
+        internal const val MAX_HEADING_LEVEL = 6
 
-        private val NUMBERED_LIST_REGEX = Regex("^(\\d+)\\.\\s*(.*)$")
-        private val IMAGE_MARKDOWN_REGEX = Regex("!\\[([^]]*)]\\(([^)]+)\\)")
-        private val RESOURCES_MARKDOWN_REGEX = Regex("!\\[[^]]*]\\((resources/[^)]+)\\)")
-        private val RESOURCES_PATH_REGEX = Regex("resources/[^/]+/[^/]+", RegexOption.IGNORE_CASE)
+        internal val NUMBERED_LIST_REGEX = Regex("^(\\d+)\\.\\s*(.*)$")
+        internal val IMAGE_MARKDOWN_REGEX = Regex("!\\[([^]]*)]\\(([^)]+)\\)")
+        internal val RESOURCES_MARKDOWN_REGEX = Regex("!\\[[^]]*]\\((resources/[^)]+)\\)")
+        internal val RESOURCES_PATH_REGEX = Regex("resources/[^/]+/[^/]+", RegexOption.IGNORE_CASE)
     }
 }
